+++ mod/app/src/main/java/bttv/api/Highlight.java
@@ -1,0 +1,9 @@
+package bttv.api;
+
+import tv.twitch.android.shared.chat.ChatMessageInterface;
+
+public class Highlight {
+    public static Integer replaceNum(ChatMessageInterface chatMessageInterface, boolean z, boolean z2, int i, int i2, Object iClickableUsernameSpanListener, Object twitchUrlSpanClickListener, Object webViewSource, String str, boolean z3, Object chatFiltersSettings, Integer num) {
+        return bttv.highlight.Highlight.replaceNum(chatMessageInterface, num);
+    }
+}