+++ mod/app/src/main/java/bttv/emote/Emotes.java
@@ -12,6 +12,7 @@ import java.util.Set;
 
 import bttv.Data;
 import bttv.Network;
+import bttv.Util;
 import bttv.settings.Settings;
 import bttv.settings.UserPreferences;
 
@@ -53,17 +54,17 @@ public class Emotes {
     public static boolean channelHasEmotes(Context context, int channelId) {
 
         // BTTV
-        boolean bttvEnabled = Settings.BTTVEmotesEnabled.entry.get(context).get();
+        boolean bttvEnabled = Util.getBooleanFromSettings(Settings.BTTVEmotesEnabled);
         boolean hasBTTVEmotes = bttvEnabled && channelEmotesBTTV.containsKey(channelId);
         boolean globalBTTVLoaded = bttvEnabled && !globalEmotesBTTV.isEmpty();
 
         // FFZ
-        boolean ffzEnabled = Settings.FFZEmotesEnabled.entry.get(context).get();
+        boolean ffzEnabled = Util.getBooleanFromSettings(Settings.FFZEmotesEnabled);
         boolean hasFFZEmotes = ffzEnabled && channelEmotesFFZ.containsKey(channelId);
         boolean globalFFZLoaded = ffzEnabled && !globalEmotes7TV.isEmpty();
 
         // 7TV
-        boolean stvEnabled = Settings.SevenTVEmotesEnabled.entry.get(context).get();
+        boolean stvEnabled = Util.getBooleanFromSettings(Settings.SevenTVEmotesEnabled);
         boolean has7TVEmotes = stvEnabled && channelEmotes7TV.containsKey(channelId);
         boolean global7TVLoaded = stvEnabled && !globalEmotes7TV.isEmpty();
 
@@ -76,10 +77,10 @@ public class Emotes {
     }
 
     public static Emote getEmote(Context ctx, String code, int channelId) {
-        boolean bttvEnabled = Settings.BTTVEmotesEnabled.entry.get(ctx).get();
-        boolean bttvGifEnabled = Settings.BTTVGifEmotesEnabled.entry.get(ctx).get();
-        boolean ffzEnabled = Settings.FFZEmotesEnabled.entry.get(ctx).get();
-        boolean stvEnabled = Settings.SevenTVEmotesEnabled.entry.get(ctx).get();
+        boolean bttvEnabled = Util.getBooleanFromSettings(Settings.BTTVEmotesEnabled);
+        boolean bttvGifEnabled = Util.getBooleanFromSettings(Settings.BTTVGifEmotesEnabled);
+        boolean ffzEnabled = Util.getBooleanFromSettings(Settings.FFZEmotesEnabled);
+        boolean stvEnabled = Util.getBooleanFromSettings(Settings.SevenTVEmotesEnabled);
 
         if (ffzEnabled) {
             Emote emote = globalEmotesFFZ.get(code);
@@ -131,9 +132,9 @@ public class Emotes {
     }
 
     public static void ensureChannelEmotes(Context context, int id) {
-        boolean bttvEnabled = Settings.BTTVEmotesEnabled.entry.get(context).get();
-        boolean ffzEnabled = Settings.FFZEmotesEnabled.entry.get(context).get();
-        boolean stvEnabled = Settings.SevenTVEmotesEnabled.entry.get(context).get();
+        boolean bttvEnabled = Util.getBooleanFromSettings(Settings.BTTVEmotesEnabled);
+        boolean ffzEnabled = Util.getBooleanFromSettings(Settings.FFZEmotesEnabled);
+        boolean stvEnabled = Util.getBooleanFromSettings(Settings.SevenTVEmotesEnabled);
 
         if (bttvEnabled && globalEmotesBTTV.isEmpty()) {
             Network.getBTTVGlobalEmotes();