+++ mod/twitch/src/main/java/tv/twitch/android/settings/base/SettingsNavigationController.java
@@ -1,0 +1,9 @@
+package tv.twitch.android.settings.base;
+
+import android.os.Bundle;
+
+import tv.twitch.android.models.settings.SettingsDestination;
+
+public interface SettingsNavigationController {
+    void navigateToSettingFragment(SettingsDestination settingsDestination, Bundle bundle);
+}