+++ mod/twitch/src/main/java/tv/twitch/android/shared/ui/menus/subscription/SubMenuModel.java
@@ -1,0 +1,9 @@
+package tv.twitch.android.shared.ui.menus.subscription;
+
+import tv.twitch.android.models.settings.SettingsDestination;
+import tv.twitch.android.shared.ui.menus.core.MenuModel;
+
+public class SubMenuModel extends MenuModel {
+    public SubMenuModel(String primaryText, String secondaryText, String auxiliaryText, SettingsDestination settingsDestination, boolean isEnabled) {
+    }
+}