+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 All notable changes to this project will be documented in this file.
 
 # master
+- adds highlight keywords feature
 
 # v0.8.0-beta
 - update base to 11.0.0